* GWorkspace/FileViewer/GWViewerBrowser.m         * GWorkspace/FileViewer/GWViewerSplit.m           more (unsigned) int to NS(U)Integer conversion
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@36075 72102866-910b-0410-8b05-ffd578937521

+++ GWorkspace/FileViewer/GWViewerBrowser.m
@@ -34,14 +34,14 @@
 
 - (id)initWithBaseNode:(FSNode *)bsnode
               inViewer:(id)vwr
-		    visibleColumns:(int)vcols 
+        visibleColumns:(int)vcols 
               scroller:(NSScroller *)scrl
             cellsIcons:(BOOL)cicns
          editableCells:(BOOL)edcells
        selectionColumn:(BOOL)selcol
 {
   self = [super initWithBaseNode: bsnode
-		              visibleColumns: vcols 
+                  visibleColumns: vcols 
                         scroller: scrl
                       cellsIcons: cicns
                    editableCells: edcells    
@@ -90,12 +90,12 @@
       NSDictionary *apps;
       NSEnumerator *app_enum;
       id key; 
-      int i;
+      NSUInteger i;
   
       if (selnodes && [selnodes count]) {
         FSNBrowserMatrix *matrix = (FSNBrowserMatrix *)[bc cmatrix];
         FSNBrowserCell *cell;
-        int row, col;
+        NSInteger row, col;
         
         location = [matrix convertPoint: location fromView: self];
